*Added Rudimentary Express code*
1. Modified some styling in searchBar.css and index.css to improve
the searchBar styling without affecting the other components.
2. Made some changes in searchBar.js to shorten the code and make it
shorter.
3. Added Express code in a new file 'server.js'. For now,
searchBar.js sends the search string to the server which responds
with the same search string. When searchBar.js receives this string,
it logs it to console (no changes visible on the UI).
5. Also added some stuff to the package.json file. Added a "proxy"
field which is used for the backend connection. Also added "express"
as a dependency
4. Will try to update README file to reflect how to run it.